+++ bigbluebutton-html5/imports/ui/components/presentation/component.jsx
