+++ bigbluebutton-html5/imports/ui/components/presentation/default-content/styles.scss
