+++ bigbluebutton-html5/imports/ui/components/presentation/service.js
