+++ bigbluebutton-html5/imports/ui/components/presentation/shape-factory/component.jsx
