+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/ellipse/component.jsx
