+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/helpers.js
