+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/line/component.jsx
