+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/pencil/component.jsx
