+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/poll/component.jsx
