+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/rectangle/component.jsx
