+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/text/component.jsx
