+++ bigbluebutton-html5/imports/ui/components/presentation/shapes/triangle/component.jsx
