+++ bigbluebutton-html5/imports/ui/components/presentation/slide-controls/service.js
