+++ bigbluebutton-html5/imports/ui/components/presentation/slide-controls/styles.scss
