+++ bigbluebutton-html5/imports/ui/components/presentation/slide/component.jsx
