+++ bigbluebutton-html5/imports/ui/components/presentation/styles.scss
